Add files via upload
Fitting routine for bond number and radius of curvature at apex of a series of experimental points. Currently just compares to known model data (taken in as "experimental points") with offset initial condition to show that the fitting routine works properly. NOTE - need to make comments and change function names.